Adds basic endpoint and ES index template tests for enabled and disabled
