+++ x-pack/test/api_integration/apis/asset_manager/index.ts
@@ -8,6 +8,6 @@ import { FtrProviderContext } from '../../ftr_provider_context';
 
 export default function ({ loadTestFile }: FtrProviderContext) {
   describe('Asset Manager API Endpoints', () => {
-    loadTestFile(require.resolve('./ping'));
+    loadTestFile(require.resolve('./basics'));
   });
 }
