--- x-pack/test/api_integration/apis/asset_manager/ping.ts
@@ -1,21 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import expect from '@kbn/expect';
-import { FtrProviderContext } from '../../ftr_provider_context';
-
-export default function ({ getService }: FtrProviderContext) {
-  const supertest = getService('supertest');
-
-  describe('basic ping endpoint', () => {
-    it('returns a successful response', async () => {
-      const response = await supertest.get('/api/asset-manager/ping').expect(200);
-
-      expect(response.body).to.eql({ message: 'Asset Manager OK' });
-    });
-  });
-}
